# api.yaml
openapi: 3.0.3
info:
  title: Placement Portal Application V2 - API Documentation
  description: |
    ## Placement Portal Application V2
    
    A comprehensive campus recruitment management system built with Flask (Backend API) and Vue.js (Frontend UI).
    
    ### Features
    - **Admin Dashboard**: Manage companies, students, and placement drives
    - **Company Portal**: Create drives, review applications, schedule interviews
    - **Student Portal**: Browse jobs, apply, track application status
    - **Automated Jobs**: Interview reminders, monthly reports, CSV exports
    - **Performance**: Redis caching for API optimization
    
    ### Authentication
    All endpoints except login/register require authentication via session cookies.
    
    ### Roles
    - `admin` - Institute placement cell (highest access)
    - `company` - Registered organizations (requires admin approval)
    - `student` - Registered students (self-registration)
    
    ### Milestones
    - Milestone 0: GitHub Repository Setup
    - Milestone 1: Database Models and Schema Setup
    - Milestone 2: Authentication and Role-Based Access
    - Milestone 3: Admin Dashboard and Management
    - Milestone 4: Company Dashboard and Job/Application Management
    - Milestone 5: Student Dashboard and Job Application System
    - Milestone 6: Job Application History and Status Tracking
    - Milestone 7: Backend Jobs (Celery + Redis)
    - Milestone 8: API Performance Optimization and Caching
    
...
  
...
servers:
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User Login
      description: Authenticate user and create session
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: User Registration
      description: Register new student or company account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid request or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User Logout
      description: End user session
      operationId: logout
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get Current User
      description: Get authenticated user profile
      operationId: getCurrentUser
      security:
        - sessionAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================================================
  # ADMIN ENDPOINTS
  # ============================================================================
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Flask session cookie for authentication
...
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
    LoginResponse:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: '#/components/schemas/User'
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
        - role
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [student, company]
        student_data:
          $ref: '#/components/schemas/StudentData'
        company_data:
          $ref: '#/components/schemas/CompanyData'
    StudentData:
      type: object
      properties:
        full_name:
          type: string
        roll_number:
          type: string
        branch:
          type: string
        year_of_study:
          type: integer
        cgpa:
          type: number
          format: float
        phone:
          type: string
        skills:
          type: string
    CompanyData:
      type: object
      properties:
        company_name:
          type: string
        industry:
          type: string
        location:
          type: string
        website:
          type: string
        hr_contact_name:
          type: string
        hr_contact_email:
          type: string
        hr_contact_phone:
          type: string
        company_description:
          type: string
    RegisterResponse:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: '#/components/schemas/User'
        redirect_url:
          type: string
    LogoutResponse:
      type: object
      properties:
        message:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, company, student]
    UserProfile:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string

    # ============================================================================
    # ADMIN SCHEMAS
    # ============================================================================
...